backend/requirements.txt
fastapi
uvicorn
torch
numpy
pillow
torch 
torchvision 
torchaudio --index-url https://download.pytorch.org/whl/cu118